Delete Image password ask logic Added

+++ Gallery_API/Routes/ImageView.js
@@ -1,26 +1,28 @@
 const router = require("express").Router()
 const Image = require("../Modals/ImageModal");
+const bcrypt = require("bcrypt")
+const User = require("../Modals/UserModal")
 
-router.post("/add",async(req,res)=>{ // for adding the New Image
+router.post("/add", async (req, res) => { // for adding the New Image
     // console.log(req.user);
-    
-    try{
-        const {label , image} = req.body
+
+    try {
+        const { label, image } = req.body
         const PostImage = new Image({
-            label:label,
-            image:image,
-            user:req.user
+            label: label,
+            image: image,
+            user: req.user
         })
-       const response = await PostImage.save()
+        const response = await PostImage.save()
         res.status(201).json({
-            status : "success",
-            data:response
+            status: "success",
+            data: response
         });
     }
-    catch(e){
+    catch (e) {
         res.status(406).json({
-            status:"Failed",
-            message:e.message,
+            status: "Failed",
+            message: e.message,
         })
     }
 })
@@ -28,7 +30,7 @@ router.post("/add",async(req,res)=>{ // for adding the New Image
 router.get("/allImage", async (req, res) => { // for getting all iMages
     // console.log(req)
     try {
-        const images = await Image.find({user: req.user}).sort({_id:-1})
+        const images = await Image.find({ user: req.user }).sort({ _id: -1 })
         res.status(200).json({
             status: "sucess",
             Images: images
@@ -37,18 +39,30 @@ router.get("/allImage", async (req, res) => { // for getting all iMages
         res.status(400).send("Failed")
     }
 });
-router.delete("/delete/:id" , async(req ,res)=>{ // for deleting the Image
-    try{
-        const _id = req.params.id
-        await Image.findByIdAndDelete(_id)
+router.delete("/delete/:id", async (req, res) => { // for deleting the Image
+    try {
+        const { password } = req.body
+        const user = await User.findOne({ _id: req.user })
 
-        return res.status(200).json({
-            message:"success"
+        bcrypt.compare(password, user.password, async function (err, result) { //Matching with Encrypted password
+            if (err) {
+                return res.status(400).send(e.message)
+            }
+            if (result) {
+                const _id = req.params.id
+                await Image.findByIdAndDelete(_id)
+
+                return res.status(200).json({
+                    message: "success"
+                })
+
+            } else {
+                return res.status(400).send("Invalid Credentials")
+            }
         })
-
-    }catch(e){
+    } catch (e) {
         return res.status(400).json({
-            "Message":e.message
+            "Message": e.message
         })
     }
 })
